moved singleton pattern to Awake() rather than Start()
This is to iron out the bug that might occurr when the MusicPlayer
starts again. It might start another instance of MusicPlayer, causing
the music to play very briefly before it gets destroyed. By using the
Awake() rather than the Start() function, since Awake() gets called
before Start() does (the latter which actually 'Starts' the music player
(hence the name)), this does not happen and the music is now
glitch-free.